// backend/src/operators/interfaces/ITour.ts
export interface ITour {
    tourLink: string;
    imageSource: string;
    title: string;
    country: string;
    resort: string;
    accommodation: string;
    duration: number;
    departureDate: string;
    arrivalDate: string;
    adultsCapacity: number;
    childrenCapacity: number;
    price: string;
}