// backend/src/operators/interfaces/ISearchParams.ts
export interface ISearchParams {
    country: string;
    resorts: string[];
    stars: string[];
    dateOfDeparture: string;
    dateOfArrival: string;
    placeOfDeparture: string;
    durationOfStay: number;
    adultsCapacity: number;
    childrenCapacity: number;
    accomodation: string[];
}